refactor(domain): dto, enums 패키지 추가
