+++ src/main/java/lotto/domain/LottoCommittee.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.dto.StatisticDto;
+
 import java.util.List;
 import java.util.stream.Collectors;
