+++ src/test/java/lotto/domain/StatisticDtoTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.dto.StatisticDto;
+import lotto.domain.enums.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
